+++ uniffi_bindgen/src/bindings/kotlin/templates/RecordTemplate.kt
@@ -1,4 +1,4 @@
-{%- let rec = ci.get_record_definition_unchecked(name) %}
+{%- let rec = ci|get_record_definition(name) %}
 
 data class {{ type_name }} (
     {%- for field in rec.fields() %}