+++ uniffi_bindgen/src/bindings/ruby/gen_ruby/mod.rs
@@ -132,6 +132,7 @@ impl<'a> RubyWrapper<'a> {
 
 mod filters {
     use super::*;
+    pub use crate::backend::filters::*;
 
     pub fn type_ffi(type_: &FfiType) -> Result<String, askama::Error> {
         Ok(match type_ {